Identity Service Implemented

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -60,12 +61,13 @@ public class AuctionsController : ControllerBase
         return _mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createdAuctionDto)
     {
         var auction = _mapper.Map<Auction>(createdAuctionDto);
-        // TODO: add current user as seller
-        auction.Seller = "test"; // test seller for user replacement
+        
+        auction.Seller = User.Identity.Name; 
         
         _context.Auctions.Add(auction);
 
@@ -83,6 +85,7 @@ public class AuctionsController : ControllerBase
         return CreatedAtAction(nameof(GetAuctionById), new {auction.Id}, newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -93,7 +96,7 @@ public class AuctionsController : ControllerBase
             return NotFound();
         }
 
-        // TODO: Check seller == username
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         // Update items if 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
@@ -112,6 +115,7 @@ public class AuctionsController : ControllerBase
         return Ok();
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -119,7 +123,7 @@ public class AuctionsController : ControllerBase
 
         if (auction == null) return NotFound("Auction could not be found.");
 
-        // TODO: check seller == username
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         _context.Auctions.Remove(auction);
